requirements.txt
Flask==0.10.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.4.0
pysqlite3==0.2.1
Werkzeug==0.14.1